requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
antiorm==1.2.1
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
db==0.1.1
db-sqlite3==0.0.1
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
greenlet==3.2.1
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.55
langchain-openai==0.3.14
langgraph==0.3.33
langgraph-checkpoint==2.0.24
langgraph-checkpoint-sqlite==2.0.6
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.63
langsmith==0.3.33
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
openai==1.76.0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
passlib==1.7.4
pdf2image==1.17.0
pillow==11.2.1
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic-extra-types==2.10.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
python-dotenv==1.1.0
python-http-client==3.3.7
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.3
rsa==4.9.1
sendgrid==6.11.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starkbank-ecdsa==2.2.0
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
xxhash==3.5.0
zstandard==0.23.0
